+++ mpf/tests/machine_files/bcp/config/config.yaml
@@ -1,0 +1,5 @@
+#config_version=4
+
+switches:
+    s_test:
+        number:
